tech-studio-projects/other-projects/multi-product-recommendation-system/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
pydantic
jinja2
pytest
pika
confluent-kafka
docker
